+++ lab/apps.py
@@ -2,5 +2,8 @@ from django.apps import AppConfig
 
 
 class LabConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'lab'
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "lab"
+
+    def ready(self):
+        from . import signals  # noqa: F401