+++ labdb/settings.py
@@ -99,7 +99,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'lab'
+    'lab.apps.LabConfig'
 ]
 
 MIDDLEWARE = [