+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+25/08/2026
+- Added entry log to the admin website.
+
+
 24/08/2026
 - Enforced uniqueness of DOB-BreedingPair-crossing-tattoo
 - Warning if trying to add a mouse where a similar mouse with the same dob,breeding pair and crossing already exists.